move tests to !netfx run
